Make glam an optional dependency
